+++ src/Entity/FooBarViewBuilder.php
