+++ src/Entity/Form/FooBarDeleteForm.php
@@ -8,6 +8,8 @@
 namespace Drupal\foo_bar\Entity\Form;
 
 use Drupal\Core\Entity\ContentEntityConfirmFormBase;
+use Drupal\Core\Form\FormStateInterface;
+use Drupal\Core\Url;
 
 /**
  * Provides a form for deleting a foo_bar entity.
@@ -25,10 +27,8 @@ class FooBarDeleteForm extends ContentEntityConfirmFormBase {
   /**
    * {@inheritdoc}
    */
-  public function getCancelRoute() {
-    return array(
-      'route_name' => 'foo_bar.list',
-    );
+  public function getCancelUrl() {
+    return new Url('foo_bar.list');
   }
 
   /**
@@ -41,11 +41,11 @@ class FooBarDeleteForm extends ContentEntityConfirmFormBase {
   /**
    * {@inheritdoc}
    */
-  public function submit(array $form, array &$form_state) {
+  public function submitForm(array &$form, FormStateInterface $form_state) {
     $this->entity->delete();
 
     watchdog('content', '@type: deleted %title.', array('@type' => $this->entity->bundle(), '%title' => $this->entity->label()));
-    $form_state['redirect_route']['route_name'] = 'foo_bar.list';
+    $form_state->setRedirect('foo_bar.list');
   }
 
 }
